pyproject.toml
[tool.poetry]
name = "sibi-scraper"
version = "0.1.4"
description = ""
authors = ["Tim Sharpe <tim@sharpe.id.au>"]
readme = "README.md"
packages = [{include = "sibi_scraper"}]

[tool.poetry.scripts]
sibi_scraper = "sibi_scraper.cli:main"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
googletrans-py = "^4.0.0"
pypdf2 = "^3.0.1"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"
ipdb = "^0.13.13"
ruff = "^0.0.291"
pdbpp = "^0.10.3"

[tool.ruff]
src = ["sibi_scraper"]
include = ["sibi_scraper/**/*.py"]
ignore = ["PLR0913"]
select = [
  "E",
  "W",
  "F",
  "UP",
  "N",
  "I",
  "YTT",
  "ASYNC",
  "S",
  "BLE",
#  "D",
#  "ANN",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "T10",
  "EM",
  "ISC",
  "ICN",
  "G",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SIM",
  "ARG",
  "PTH",
  "TD",
  "FIX",
  "ERA",
  "PL",
  "TRY",
  "PERF",
  "RUF",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
